requirements.txt
flask
scrapetube
requests
google-generativeai
python-dotenv
